requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.3.5
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
numpy>=1.24.3
pandas>=2.0.3
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6 